cmd/juju: add --force to destroy-environment
Also, fix add-machine (don't convert nil *params.Error
to error), to stop the spurious error message.

+++ cmd/juju/destroyenvironment.go
@@ -16,8 +16,8 @@ import (
 	"launchpad.net/juju-core/environs"
 	"launchpad.net/juju-core/environs/configstore"
 	"launchpad.net/juju-core/juju"
-	"launchpad.net/juju-core/rpc"
 	"launchpad.net/juju-core/state/api"
+	"launchpad.net/juju-core/state/api/params"
 )
 
 var NoEnvironmentError = errors.New("no environment specified")
@@ -27,6 +27,7 @@ type DestroyEnvironmentCommand struct {
 	cmd.CommandBase
 	envName   string
 	assumeYes bool
+	force     bool
 }
 
 func (c *DestroyEnvironmentCommand) Info() *cmd.Info {
@@ -40,6 +41,7 @@ func (c *DestroyEnvironmentCommand) Info() *cmd.Info {
 func (c *DestroyEnvironmentCommand) SetFlags(f *gnuflag.FlagSet) {
 	f.BoolVar(&c.assumeYes, "y", false, "Do not ask for confirmation")
 	f.BoolVar(&c.assumeYes, "yes", false, "")
+	f.BoolVar(&c.force, "force", false, "Forcefully destroy the environment, directly through the environment provider")
 }
 
 func (c *DestroyEnvironmentCommand) Run(ctx *cmd.Context) error {
@@ -65,14 +67,19 @@ func (c *DestroyEnvironmentCommand) Run(ctx *cmd.Context) error {
 			return errors.New("Environment destruction aborted")
 		}
 	}
-	conn, err := juju.NewAPIConn(environ, api.DefaultDialOpts())
-	if err != nil {
-		return err
-	}
-	defer conn.Close()
-	err = conn.State.Client().DestroyEnvironment()
-	if err != nil && !rpc.IsNoSuchRequest(err) {
-		return fmt.Errorf("could not remove agents: %v", err)
+	// If --force is supplied, then don't attempt to use the API.
+	// This is necessary to destroy broken environments, where the
+	// API server is inaccessible or faulty.
+	if !c.force {
+		conn, err := juju.NewAPIConn(environ, api.DefaultDialOpts())
+		if err != nil {
+			return err
+		}
+		defer conn.Close()
+		err = conn.State.Client().DestroyEnvironment()
+		if err != nil && !params.IsCodeNotImplemented(err) {
+			return fmt.Errorf("destroying environment: %v", err)
+		}
 	}
 	return environs.Destroy(environ, store)
 }
